requirements.txt
scipy
trimesh
libigl
matplotlib
pybind11
#torch_sparse_solve
tqdm
omegaconf
ninja
imageio
imageio-ffmpeg
rich
diffusers
transformers
accelerate
jaxtyping
html4vision
netvision
kornia
triangle